mydjango04/requirements.txt
django~=4.2.7
django-bootstrap5==23.3
django-debug-toolbar==4.2.0
# hottrack app
requests==2.31.0
pillow==10.1.0
pandas==2.1.3
django-extensions~=3.2.3
pymysql==1.1.0
cryptography==41.0.7
psycopg2-binary==2.9.9
Werkzeug==3.0.1
django-vanilla-views==3.0.0
django-environ==0.11.2
django-widget-tweaks~=1.5.0
django-crispy-forms~=2.1
crispy-bootstrap5~=2023.10